[DevTools] Include issueSummary in exceptionMetadata for CORS errors.
Gated behind the existing `DevToolsImprovedNetworkError` feature, this
adds additional metadata to every CORS error object thrown by `fetch()`.
This metadata is inaccessible to the page itself and propagated via a
side channel to the inspector.

The issue summary contains a brief summary of the CORS error (caused
by `fetch()`) with some hints for the developer how to fix it. The full
information can still be accessed in the Issues panel if necessary.

The DevTools front-end also needs to be updated (in a follow up step)
to consume the issue summary and present it to the developer as part
of the console message that is generated for the uncaught exception.

As part of this CL, we also prevent the generation of the detailed
but redundant console message for CORS errors in the `ResourceLoader`
(again gated behind the `DevToolsImprovedNetworkError` feature) for
the case of `fetch()`. We will need to do something similar for XHR
later.

Doc: https://goo.gle/devtools-reduce-network-noise-design
Bug: 371523542, 371503691
Change-Id: I79fc9858684f88391eff4de361cc33a96545a473
Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5999257
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381133}
NOKEYCHECK=True
GitOrigin-RevId: 544d046580c3d4b2ea33fff5d5f6b20e900abcad