Fix pointerup events leaking from the overlay into the inspected page
This CL fixes an issue where an inspected page would see 'pointerup'/
'mouseup' events after the developer clicked the "Resume" button in
the DevTools overlay.

The issue is that the overlay synchronously resumes the underlying
V8 isolate from the overlay's 'click' handler. So while we are in the
middle of processing/propagating the "MouseReleased" event, we are
already resuming the inspected frame.

This caused the "MouseReleased" event to show up as a 'pointerup'
DOM event in the inspected page if the area underneath the debugger
toolbar has a corresponding handler installed.

The solution is to delay resuming the overlay until we are done
with the "MouseReleased" event.

We have to land the regression test disabled because the
"ClickToCapturedPointer" experiment doesn't work with the overlay
at the moment. We'll enable the test once the experiment/overlay
interaction is fixed.

R=alexrudenko@chromium.org

Bug: 378451943
Change-Id: I69cdc52f6e8288ea7bf32c202d93144218472ede
Fixed: 40833742
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6010733
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381152}
NOKEYCHECK=True
GitOrigin-RevId: c803a005d736e1108ed2a4d8e5454c969f412822