Add base_subdirs to file_chooser.mojom NativeFileInfo
Android content-URIs do not contain the names of subdirs like paths in
posix or windows do.

This CL is a nop to add the field to mojom and update any affected
clients. It is part of a series of CLs to get webkitRelativePath
working for android content-URIs.

Bug: 377716464
Change-Id: Idaa6de86d81a705c4d7758929b34ea5ef2101a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6000058
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381493}
NOKEYCHECK=True
GitOrigin-RevId: 0174ba71d0cd4719e9dda16d28c7181c29db9157