[bfcache]Update reason names to match the spec
This CL updates the reason names reported by NotRestoredReasonsAPI so
that they match the ones in the spec.
On the spec PR we decided some reasons should not be exposed, i.e.
reported as masked, and changed other reasons names. Mainly we are
changing the reporting name to be 'masked' in this CL.
The change is guarded with a runtime flag.

Note that session restore should be reported as null but with this CL it
is still reported as "session-restored". I will address this case and
add a test in a separate CL.

Spec PR: https://github.com/whatwg/html/pull/10154

PSA: https://groups.google.com/a/chromium.org/g/blink-dev/c/keA8QE_uFxo/m/OqtfRXq6BwAJ?utm_medium=email&utm_source=footer
Bug: 331754704
Change-Id: I0d4224f599bc4cdae7a70b9bbb38ff798bae2917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5907581
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381607}
NOKEYCHECK=True
GitOrigin-RevId: 5dc27faa601e026975511cfb1c489c753d2cc038