Prompt API: leave the sampling params null if the model is not ready
Since the sampling parameters are configured from the model execution
config, it's not possible to get the value with the model being
downloaded. We will leave those fields in the `capabilities()` API null
for the "after-download" case just like the "no" case.

Bug: 377923490
Change-Id: I1b9ce841d452048f1b8838927903f8c7dbacec9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012500
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381631}
NOKEYCHECK=True
GitOrigin-RevId: 64479fd017bdffe0e9e6de4ad4c2a3c8643db5ca