[Sanitizer] Implement Builtins (like the default config(s)).
This implements the Sanitizer builtins, using code generators so that
we can store them as QualifiedNames. This ought to safe a good bit of
memory (vs naively storing them as text, or name/namespace pairs). It
should also make it easier to adapt the configs as the specification
evolves.

A rough guide to the implementation:
- builtins_*.json: These are the "real" sources for the builtins.
- generate_*.py: These are scripts to generate C++ constants from the
  sources.
- sanitizer_builtins.h/.cc: The API to access the builtins.
- sanitizer_bulitins_unittest.cc: Sanity checks for the builtins, so that
  future me will not inadvertently violate the Sanitizer guarantees.

One generated file (builtins_all_known.cc) is quite large and is only
used for testing. We need to take some care in the buildfiles to make sure
it will not get included in production binaries.

Bug: 356601280
Change-Id: I91660bbe8bd4265fa3cb607a0da622179d81be53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5993856
Reviewed-by: Yifan Luo <lyf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381817}
NOKEYCHECK=True
GitOrigin-RevId: f92385a4f93ba7274a015880b8ad4d0ba51971d6