[Gap Decorations]: Generalize `column-rule-color` parsing code
This CL generalizes the parsing logic for `column-rule-color` by
introducing the `ConsumeGapDecorationPropertyValue` function. This
function consumes different values based on the
`CSSGapDecorationPropertyType`. Following this, calls to `ConsumeColor`
are replaced with `ConsumeGapDecorationPropertyValue`. This approach
avoids the need to repeat the entire parsing logic when the only
difference is the value consumed.

Bug: 357648037
Change-Id: Ibfb47e9f1ad744d6df58917b70d576f79034ef65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5990251
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1381901}
NOKEYCHECK=True
GitOrigin-RevId: e1dc1b2948d57df002063cec013a290bdedd495f