[Headless]: Fix Forced Colors Tests for Headless Shell
This CL resolves issues with the forced colors mode tests in the
headless shell test environment. The errors occurred because the link
color in some reference files was not explicitly set to "VisitedText",
as is currently done in forced colors mode. Consequently, many actual
results did not match the expected results solely due to the link color
discrepancy. For more details, refer to this test bucket[1].

Most errors in the Windows scenario should be resolved with this
change, except for `forced-colors-mode-03.html`, which requires further
restructuring based on recent findings about the test's structure. This
will be addressed in a subsequent CL.

Failures on MacOS and Linux platforms will also be handled in a
subsequent CL.

[1]: https://chromium-layout-test-archives.storage.googleapis.com/results.html?json=chromium/ci/win10-wpt-chromium-rel/3919/headless_shell_wpt_tests_include_all%20%28experimental%29/full_results_jsonp.js

Bug: 368396826
Change-Id: Ifc925637e0f1438a5241017de7a2f3023380997e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012900
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1382010}
NOKEYCHECK=True
GitOrigin-RevId: 6782f0ae4c9dd21ae215edda569b2a241ed55ef7