[@scope] Make selector parsing @scope-agnostic
The selector parser currently knows whether it's within @scope
or not, and (if applicable) emits kScopeActivation relations
before any compound which contains ':scope' or '&'. SelectorChecker
then later uses this relation to actually calculate the scoping roots.

Unfortunately, this approach is not very compatible with the planned
scoped import feature, e.g: @import url("foo.css") scope(.mycomponent).
In this case, the stylesheet (foo.css) is parsed without knowing
that it's actually scoped. Additionally, the same stylesheet may be
imported in multiple places, with different scoping applied to each
case.

To make this feature implementable in a civilized manner, this CL
removes the kScopeActivation relation, instead marks any CSSSelector
with a flag (IsScopeContaining), and "activates" scopes (when needed)
at the start of SelectorChecker::MatchSelector for any selectors
with this flag set.

Note: Proximity propagation in ~SubResult is no longer needed,
      because a deeply nested '&' rule will activate scopes now.
      It's possible that this could have been removed earlier, i.e.
      I suspect some other CL already made this code obsolete.

Note: The kPseudoScope "fast path" in the kDescendant branch of
      SelectorChecker::MatchForRelation is not correct when :scope
      does not represent a tree-scope, so a !context.style_scope
      condition was added. The implicitly inserted :true selector
      avoided hitting this problem earlier, but that was not intentional.
      This path is suspected of being wrong altogether, tracked by
      Issue 378698644.

Bug: 369876911, 378698644
Change-Id: I1f352c98309a1e4b511841df48f497678fd7f72e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6015378
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382306}
NOKEYCHECK=True
GitOrigin-RevId: 9652a5bae1ade7852b1f9594e4b8671d02d5bce9