Map snap-event WPTs to web-features
web-features[1] is an effort to build a shared catalog of features of the web platform.

This patch adds some of the necessary configuration to ensure that
snap events are represented in web-features. Snap Events have been
added[2] to the web-features repository as
scroll-snap-events.

[1] https://github.com/web-platform-dx/web-features
[2] https://github.com/web-platform-dx/web-features/commit/e4c58050d0a65e53050cbdd0119ee87f423b8690

Bug: 365784987
Change-Id: I31c901305dd72fde21a5d5cf1176604fbaa91bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5851064
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382338}
NOKEYCHECK=True
GitOrigin-RevId: 1741f27a85b853336a6b8bdf553487ad763d7cb7