Make ScriptPromise STACK_ALLOCATED
In the vast majority of usage, ScriptPromise is only ever on the
stack. Add a new variant, StoredScriptPromise, for the few cases
where the ScriptPromise does need to be stored. The
StoreScriptPromise only provides based operations like IsEmpty(),
but retains the underlying v8::Promise and the typing information
so that it can be converted back to a stack allocated ScriptPromise
when needed.

This allows us to avoid GC opeartions in the most common usage of
ScriptPromise, minimizing the overhead of ScriptPromise relative to
a v8::Promise.

Change-Id: I9dc32eb03dbbfae2c8eff7d2534dacc5603d9a78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6013586
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382642}
NOKEYCHECK=True
GitOrigin-RevId: 3935a121ea60c03f1a90348ce7e1c58d34776c8b