Prompt API: temporarily allow timeout for the context overflow WPT
When the prompt request doesn't get a response (in the case reported
from 378549335, it's caused by the requested being filtered), the new
prompt won't be put into the context. In the context overflow WPT, we
has a loop to keep prompting the model, expecting the context will
eventually overflow, but if the model always throws errors, the loop
will never end, which causes the timeout.

This test is only disabled on Mac because only optimization_guide-mac-*
runs the test with the real model.

Bug: 378549335
Change-Id: Iea1eb4fe879c5d3f93c1fa209c75be0b45f91dcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6013787
Reviewed-by: Fergal Daly <fergal@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382860}
NOKEYCHECK=True
GitOrigin-RevId: 19ebcba44104976c682f12f66ed6ebce746b3533