spanification: Spanify WTF::unicode::ConvertLatin1ToUTF8
Add a new version of said function that takes one span for the input
(Latin-1 characters) and one span for the output (UTF-8 code units). For
now this function only wraps the old version. Instead of using out-
parameters, return a struct with a span for part of the output that
contains converted code units, the number of characters consumed from
the input and the status (previous result).

Bug: 351564777
Change-Id: I5004e41940203162791645b3d5589d5a64357272
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6018210
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382909}
NOKEYCHECK=True
GitOrigin-RevId: 3580a514d3d4146237ae97598a98bc25dc86ce24