Disable CSS Highlight Inheritance (again)
This needs to be rolled out due to breakage only noticed
in stable.

Most likely this was not noticed in previous attempts to
launch due to a recent update in Tailwind CSS which defines
custom properties inside a ::selection, even when they
are not necessary. Or maybe it's always been that way.

Two paths forward for launching this again: Handle the use
case of custom properties defined and used within a single
::selection block; or fix Tailwind CSS and wait for it to
propagate. Probably need the former.

Bug: 1024156
Change-Id: I01e67319239d5ab2647666a18c4e3aa20b930ef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6023765
Auto-Submit: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1383110}
NOKEYCHECK=True
GitOrigin-RevId: 921435f584725f0d3da0b27e089af2fcc41686a5