[Headless]: Update `forced-colors-mode-03` to match expected behavior
This CL updates the `forced-colors-mode-03.html` to align with the
correct behavior. Initially, the test was structured in a way that did
not accurately reflect its semantic representation. Specifically, it
included two anchor tags (with ids `h` and `i`) nested in another
anchor tag (with id `g`). This created a false impression of a
parent-child relationship existing between anchor tag `g` and anchor
tags `h` and `i`. In reality, this relationship does not exist and the
parent of anchor tags `h` and `i` is the `body` and not anchor tag `g`.

Additionally, the test did not account for the fact that
`getComputedStyle()` returns the unvisited link color on anchor tags
for privacy reasons. This CL addresses these issues by updating the
test's structure and expectations accordingly.

With this CL, all forced-colors-mode/ tests for the headless test
environment are fixed on Windows. Failures on MacOS and Linux platforms
will also be handled in a subsequent CL.

Bug: 368396826
Change-Id: I3e3b10651c2fb3ed3edcfd2009170c8e34af5a7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6015699
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1383262}
NOKEYCHECK=True
GitOrigin-RevId: 98303b2ad523bc059d647a67e08257b72b8b6da0