[A11y] Expose presence of potential interest target
An interesttarget is an experimental HTML attribute that, along with
an interestaction, allows for the target's visibility to be toggled
on mouse hover. However, an AT user must be informed that there
is potentially an object there.

- Expose a new object attribute for "rich" interest targets that are not just plain text (plain text targets are exposed as descriptions)
- Unlike the details relation, the state is exposed even if if the
target is currently hidden

This will allow ATs to inform users that there is an interest target
present, so that they can use the command to bring it up
(e.g. ctrl+insert+enter  should work on JAWS, which routes the
mouse cursor to the object to trigger any mouseover effects).

Fixed: 377923492
Change-Id: I6103ce77154e94b14d67f6195888efc616e220f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012299
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1383299}
NOKEYCHECK=True
GitOrigin-RevId: d62ef43218b7401df20088ea10937abdd1a79a09