ipcz: Fix various potential driver object leaks
This introduces a bit of extra caution in cleaning up dangling object
references (e.g. through retained callbacks) which could lead to leaks
when internal communications are tampered with.

Also corrects a few bits of code which could assert in debug builds
(but which were effectively safe in production) under such conditions.

This is a precursor to landing a fuzzer for ipcz.

Bug: 370880241
Change-Id: I03455475adf3401ada6fb81ba192fa709156f83f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6019406
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1383331}
NOKEYCHECK=True
GitOrigin-RevId: 200eab2fa0a0a74b6472b709a1e6677c1b969fb6