Refactor pointerevent_poitnerId_scope test
This test was periodically flaking on debug builds.  A timeout would
resolve if we did not see the expected number of events. This is
possible due to event coalescing. Added a click on each of the targets
to ensure that we get the expected number of events.

Also removed the manual testing instructions and converted the test to
be promise driven.

Bug: 352327319
Change-Id: I3aefac84b6be0ca76a9678e03eef3c89d0e18092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020360
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384414}
NOKEYCHECK=True
GitOrigin-RevId: a9a6013ba4636e56ed5326cc10f361fd46c8e603