Fix clipboardevent findtarget logic to use focused element
As per the clipboard event spec, to find the target element for
clipboard events like "onPaste", "onCut", etc, if the current element
is not editable, then we set the currently focused element as the target
otherwise we set the document body as the target. Currently, if there is
no visible selection, we don't consider focused element.

Added web test for this scenario and fixed an earlier web test where the
focused element was not considered as the target of clipboard events.

Spec: https://www.w3.org/TR/clipboard-apis/#fire-a-clipboard-event

Bug: 40735783
Change-Id: I2cab64c7e4be6c862e5cd7718b82071dd5e05488
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6011047
Reviewed-by: Siye Liu <siliu@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Rohan Raja <roraja@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1384755}
NOKEYCHECK=True
GitOrigin-RevId: ce799d15c8ba46567db41e686bdff29bb2c217be