Truly fix scroll-marker updates for programmatic scrolls
The original cl[1] failed to recognize that scroll updates from the
compositor might be in service of an ongoing animation and just used
the scroller's current scroll offset[2]. This cl fixes this so that
ongoing scroll animations are always accounted for.

This patch makes a few style changes to the scroll container in the
test that added for this issue
(scroll-markers-updated-by-programmatic-scroll.html) so that scroll
animations on it can be composited.
However, note that it appears that under the current configurations
under which the test is run (i.e. no virtual suites[3], so no threaded
animations) the issue would not surface. So, the test is added
under the virtual/threaded test suite in this patch.

[1] https://crrev.com/c/5985230
[2] https://source.chromium.org/chromium/chromium/src/+/39675f74ca0c146e7c3e1d0a82ade79631b9cef7:third_party/blink/renderer/core/frame/web_frame_widget_impl.cc;l=1485
[3] https://ci.chromium.org/ui/p/chromium/test-search?q=wpt_internal%2Fcss%2Fcss-overflow%2Fscroll-markers-updated-by-programmatic-scroll.html

Bug: http://crbug.com/377321461
Change-Id: I0a91b5a411fbe06972c09047a7fd6007e85abd09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6019895
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385033}
NOKEYCHECK=True
GitOrigin-RevId: 2484486a93bbf2f8fb9459d6e079b16bb4a18240