DOM: Flag-guard moveBefore()'s range and selection preservation
Per discussion in https://github.com/whatwg/dom/pull/1307, we've
decided to not ship range and selection preservation initially. Since
this was mostly implemented in Chromium, this CL flag-gaurds our
implementation of that preservation (so it does not ship along with
the rest of the API's side-effects) and updates the tests.

R=nrosenthal@chromium.org

Bug: 40150299
Change-Id: Ia4412d95859497593ac2e4d9e9b87dfc36240ef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6022661
Reviewed-by: Noam Rosenthal <nrosenthal@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385070}
NOKEYCHECK=True
GitOrigin-RevId: 3bececa95583b2ff0226e430c21d30c0f2b63ac6