Don't pass null expression from InterpolableNumber::Scale.
This avoids passing a null expression in InterpolableNumber::Scale(const
InterpolableNumber&), which was previously possible when
this->IsDoubleValue() but the earlier branch is not taken.  It makes the
same change to InterpolableNumber::Scale(double) even though that change
is not needed today, to reduce risk of the same problem happening there
as a result of future changes.

This fixes a regression from
https://crrev.com/93eaf2b2bfc793353bd4a86ede01e9b0e2bfaf8d .

Fixed: 375798131
Change-Id: I98e17ab8afca29fa6924585eb77cc496290bfa37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6034421
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385252}
NOKEYCHECK=True
GitOrigin-RevId: b01e368d30a339041b93ce4ddf3c232099d02a74