WebGPU: Add 1-component GPUVertexFormats and unorm8x4-bgra
This adds a runtime_enabled_feature for feature gating and performs a
validation pass after conversion of the GPURenderPipelineDescriptor to
the wgpu::RenderPipelineDescriptor to check if any of the new formats
are used without the feature enabled. This check will be removed once
the feature is safely landed.

Bug: 376924407
Change-Id: Ib1b51b01cbfa72da3c20f8519cd9b06865bfc817
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6032609
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Fernando Serboncini <fserb@chromium.org>
Auto-Submit: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385257}
NOKEYCHECK=True
GitOrigin-RevId: 2b70c1ff902f60d5580637388eae0f2d5e5f0fda