Correct additional test for inlines in item-less container fragments.
The two tests in crrev.com/c/6022897 were identical, except for the
description. I had prepared two different tests, but somehow only
the description made it past the finishing line, and not the actual
test contents.

Thanks to dholbert for noticing this it in
https://github.com/web-platform-tests/wpt/pull/49174

Bug: 40287080
Change-Id: I5e0e078c5a7780b02c47113a1b51d86868479bbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6034400
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385266}
NOKEYCHECK=True
GitOrigin-RevId: b13e7a3ef41538e86e5eebe3aeb198bd1933834d