Use ThenCallable in PipeToEngine
The overhead of ScriptPromise relative to v8::Promise is now low
enough that it makes sense to use ScriptPromise all the way
through the streams/ plumbing, and remove the streams/ specific
promise helpers. This does that conversion for PipeEngine, and
makes it ~15% faster in microbenchmark.

Change-Id: I786fee78bd3b63a31846a4ba91ce6c4e8cc826fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980394
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385296}
NOKEYCHECK=True
GitOrigin-RevId: e7a171e5b4b33e90033bd214f318b31c96f944c5