Slightly modify slot assignment invalidation fix
The fix for issue 40127208[1] was not properly documented, so added
documentation to clarify what this invalidation is about.

It was unclear in the original fix whether the computed style check
for assigned node is strictly necessary or even an optimization. Kept
it for now, but rewrote to not use
GetComputedStyleForElementOrLayoutObject().

TEST=wpt/html/rendering/replaced-elements/the-select-element/select-multiple-re-add-option-via-document-fragment.html

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2160930

Bug: 40286814
Change-Id: I8fc1ce3370d1e40ac22fa53e2972a8ea87f23ca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5937928
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385554}
NOKEYCHECK=True
GitOrigin-RevId: a17e0ea7bd77e029f43ce1ed41ad917d59b76d88