Fix a crash in -webkit-gradient() parsing.
This happens when color-stop() has calc() expressions that are not
possible to evaluate before the computed style is evaluated.
The fix is trivial, though, as we already have the machinery in place
to deal with such values (possibly due to other gradient types?).

Bug: 40946458
Change-Id: Idf28e9880694564cc3999f9bf927c557a3010c9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6033606
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1385575}
NOKEYCHECK=True
GitOrigin-RevId: 9eb167e55f312669f5f997efe7c07dbe541a2c85