[Blink] Rationalize CanvasRC2D::CanCreateCanvas2DResourceProvider()
Changes this method to try to create a CanvasResourceProvider and return
whether that succeeded whether than trying to create a
Canvas2DLayerBridge and return whether that succeeded. The former is the
intent of this method, which was fulfilled back when Canvas2DLayerBridge
held the resource provider (or its equivalent). However, that has not
been the case for a long time.

As this is a behavioral change, we put it under a killswitch for safety.

Change-Id: Ia888a88faeafa288b5977f6fe07c4a51495f4d8d
Bug: 40280152
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6037932
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1386089}
NOKEYCHECK=True
GitOrigin-RevId: 987b74451cc7cefc1e1a7b889feb4f0cb3683697