spanification: StringImpl::Replace(pos, len, replacement)
Factor out one helper for computing the size of the string with the
replacements applied, and one helper for copying from a StringView to a
UChar span. These helpers will be used when converting some of the other
Replace() implementations as well.

Bug: 351564777
Change-Id: I8df919f13166ac48ad8b5300fac7dec8d7f85fd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6038292
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1386151}
NOKEYCHECK=True
GitOrigin-RevId: 74f4bf82f9816749b5033e1df1954af821ee4b2f