[Blink] Eliminate checks on context provider wrapper's context provider
WebGraphicsContext3DProviderWrapper::ContextProvider() always returns a
valid pointer [1]. However, a handful of the 100+ callsites of this
method check for whether the pointer is null. This CL removes those
unnecessary checks that I missed in the first sweep [2]. A followup will
change this method to return a reference.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6031880
[2] https://chromium-review.googlesource.com/c/chromium/src/+/6037225

Bug: 40280152
Change-Id: Ic2c3c24df1083da32331eb2f7b9c0e488a745896
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6037706
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1386154}
NOKEYCHECK=True
GitOrigin-RevId: 2a83b99ad4cfddaf175870a1f73db3c6de69448f